Flatten HookService into WebhookService
